+++ modules/fireback/codegen/react-new/src/modules/fireback/components/reactive-search/useReactiveSearchOld.tsx
@@ -1,0 +1,111 @@
+// I have copied this code, from older generated code.
+// it needs to be deleted, and new emi generated code has to be used.
+
+/*
+*	Generated by fireback 1.2.5
+*	Written by Ali Torabi.
+* The code is generated for react-query@v3.39.3
+*	Checkout the repository for licenses and contribution: https://github.com/torabian/fireback
+*/
+import { type FormikHelpers } from "formik";
+import React, {
+    useCallback,
+    useContext,
+    useState,
+    useRef,
+    useEffect
+} from "react";
+import {
+    useMutation,
+    useQuery,
+    useQueryClient,
+    type QueryClient,
+    type UseQueryOptions
+} from "react-query";
+import { RemoteQueryContext } from "../../sdk/core/react-tools";
+
+interface ReactiveQueryProps {
+    query?: any,
+    queryClient?: QueryClient,
+    unauthorized?: boolean,
+    execFnOverride?: any,
+    queryOptions?: UseQueryOptions<any>,
+    onMessage?: (msg: string) => void;
+    presistResult?: boolean;
+}
+export function useReactivereactiveSearch({
+    queryOptions,
+    execFnOverride,
+    query,
+    queryClient,
+    unauthorized,
+    onMessage,
+    presistResult,
+}: ReactiveQueryProps) {
+    const { options } = useContext(RemoteQueryContext);
+    const remote = options.prefix;
+    const token = options.headers?.authorization;
+    const workspaceId = (options.headers as any)["workspace-id"];
+    const connection = useRef<WebSocket>();
+    const [result, setResult] = useState<any[]>([]);
+    const appendResult = (result: any) => {
+        setResult((v) => [...v, result]);
+    };
+    const [connected, setConnected] = useState(false);
+    const close = () => {
+        if (connection.current?.readyState === 1) {
+            connection.current?.close();
+        }
+        setConnected(false);
+    };
+    const write = (data: string | Blob | BufferSource) => {
+        connection.current?.send(data);
+    };
+    /*
+    * Creates the connection and tries to establish the connection
+    */
+    const operate = (value: any, callback: any = null) => {
+        if (connection.current?.readyState === 1) {
+            connection.current?.close();
+        }
+        setResult([]);
+        const wsRemote = remote?.replace("https", "wss").replace("http", "ws");
+        const remoteUrl = `/reactive-search`.substr(1)
+        let url = `${wsRemote}${remoteUrl}?acceptLanguage=${(options as any).headers["accept-language"]
+            }&token=${token}&workspaceId=${workspaceId}&${new URLSearchParams(
+                value
+            )}&${new URLSearchParams(query || {})}`;
+        url = url.replace(":uniqueId", query?.uniqueId);
+        let conn = new WebSocket(url);
+        connection.current = conn;
+        conn.onopen = function () {
+            setConnected(true);
+        };
+        conn.onmessage = function (evt: any) {
+            if (callback !== null) {
+                return callback(evt);
+            }
+            if (evt.data instanceof Blob || evt.data instanceof ArrayBuffer) {
+                onMessage?.(evt.data);
+            } else {
+                try {
+                    const msg = JSON.parse(evt.data);
+                    if (msg) {
+                        onMessage && onMessage(msg);
+                        if (presistResult !== false) {
+                            appendResult(msg);
+                        }
+                    }
+                } catch (e: any) {
+                    // Intenrionnaly left blank
+                }
+            }
+        };
+    };
+    useEffect(() => {
+        return () => {
+            close();
+        };
+    }, []);
+    return { operate, data: result, close, connected, write };
+}
